Refactor and move test
