+++ source/Processing.Tests/Domain/BusinessProcesses/MoveIn/ConsumerMoveInTests.cs
@@ -18,6 +18,8 @@ using Processing.Domain.Consumers;
 using Processing.Domain.EnergySuppliers;
 using Processing.Domain.MeteringPoints;
 using Processing.Domain.MeteringPoints.Events;
+using Processing.Domain.MeteringPoints.Rules.ChangeEnergySupplier;
+using Processing.Domain.SeedWork;
 using Xunit;
 
 namespace Processing.Tests.Domain.BusinessProcesses.MoveIn;
@@ -27,29 +29,56 @@ public class ConsumerMoveInTests
     private readonly AccountingPoint _accountingPoint;
     private readonly Consumer _consumer;
     private readonly EnergySupplier _energySupplier;
+    private readonly ISystemDateTimeProvider _systemDateTimeProvider;
+    private readonly ConsumerMoveIn _consumerMoveIn;
+    private readonly Transaction _transaction;
 
     public ConsumerMoveInTests()
     {
+        _consumerMoveIn = new ConsumerMoveIn();
+        _systemDateTimeProvider = new SystemDateTimeProviderStub();
         _accountingPoint = AccountingPoint.CreateProduction(GsrnNumber.Create(SampleData.GsrnNumber), true);
         _consumer = new Consumer(ConsumerId.New(), CprNumber.Create(SampleData.ConsumerSocialSecurityNumber), ConsumerName.Create(SampleData.ConsumerName));
         _energySupplier = new EnergySupplier(EnergySupplierId.New(), GlnNumber.Create(SampleData.GlnNumber));
+        _transaction = Transaction.Create(SampleData.Transaction);
     }
 
     [Fact]
     public void Throw_if_any_business_rules_are_broken()
     {
-        var consumerMoveIn = new ConsumerMoveIn();
-        consumerMoveIn.StartProcess(_accountingPoint, _consumer, _energySupplier, SystemClock.Instance.GetCurrentInstant(), Transaction.Create(SampleData.Transaction));
+        _consumerMoveIn.StartProcess(_accountingPoint, _consumer, _energySupplier, SystemClock.Instance.GetCurrentInstant(), _transaction);
 
-        Assert.Throws<BusinessProcessException>(() => consumerMoveIn.StartProcess(_accountingPoint, _consumer, _energySupplier, SystemClock.Instance.GetCurrentInstant(), Transaction.Create(SampleData.Transaction)));
+        Assert.Throws<BusinessProcessException>(() => _consumerMoveIn.StartProcess(_accountingPoint, _consumer, _energySupplier, SystemClock.Instance.GetCurrentInstant(), _transaction));
     }
 
     [Fact]
     public void Consumer_move_in_is_accepted()
     {
-        var consumerMoveIn = new ConsumerMoveIn();
-        consumerMoveIn.StartProcess(_accountingPoint, _consumer, _energySupplier, SystemClock.Instance.GetCurrentInstant(), Transaction.Create(SampleData.Transaction));
+        _consumerMoveIn.StartProcess(_accountingPoint, _consumer, _energySupplier, SystemClock.Instance.GetCurrentInstant(), _transaction);
 
         Assert.Contains(_accountingPoint.DomainEvents, e => e is ConsumerMoveInAccepted);
+    }
+
+    [Fact]
+    public void Cannot_move_in_on_a_date_where_a_move_in_is_already_registered()
+    {
+        var moveInDate = _systemDateTimeProvider.Now();
+
+        _consumerMoveIn.StartProcess(_accountingPoint, _consumer, _energySupplier, moveInDate, _transaction);
+
+        var result = _consumerMoveIn.CheckRules(_accountingPoint, moveInDate);
+        Assert.Contains(result.Errors, error => error is MoveInRegisteredOnSameDateIsNotAllowedRuleError);
+    }
+
+    [Fact]
+    public void Should_return_error_when_a_completed_movein_with_the_same_movein_date_exists()
+    {
+        var moveInDate = _systemDateTimeProvider.Now();
+        _accountingPoint.AcceptConsumerMoveIn(_consumer.ConsumerId, _energySupplier.EnergySupplierId, moveInDate, _transaction);
+        _accountingPoint.EffectuateConsumerMoveIn(_transaction, _systemDateTimeProvider);
+
+        var result = _accountingPoint.ConsumerMoveInAcceptable(moveInDate);
+
+        Assert.Contains(result.Errors, error => error is MoveInRegisteredOnSameDateIsNotAllowedRuleError);
     }
 }
